Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
openai==1.35.13
pdfplumber==0.11.2
qdrant-client==1.10.1
redis==5.0.7
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
tenacity==8.4.1
structlog==24.2.0
numpy==1.26.4
